% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_suggested_pkg}
\alias{check_suggested_pkg}
\title{Check that a suggested package is available}
\usage{
check_suggested_pkg(pkg)
}
\arguments{
\item{pkg}{Package name as a string.}
}
\value{
\code{TRUE} invisibly if available; throws an error otherwise.
}
\description{
Check that a suggested package is available
}
\keyword{internal}
