% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{stac_registry}
\alias{stac_registry}
\title{Built-in catalog registry}
\usage{
stac_registry()
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns: \code{name}, \code{url}, \code{provider}.
}
\description{
Returns a tibble of known STAC API endpoints bundled with the package.
}
\keyword{internal}
