% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacr-package.R
\docType{package}
\name{stacr-package}
\alias{stacr}
\alias{stacr-package}
\title{stacr: Tidy STAC Workflows for R}
\description{
Wraps the 'rstac' package with a pipe-friendly, tidy API. All results return tibbles instead of nested lists. Ships with a catalog registry of known STAC endpoints including Planetary Computer, Earth Search, and USGS, while supporting any STAC API URL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/null-island-labs/stacr}
  \item \url{https://null-island-labs.github.io/stacr/}
  \item Report bugs at \url{https://github.com/null-island-labs/stacr/issues}
}

}
\author{
\strong{Maintainer}: Chris Lyons \email{chris@nullislandlabs.dev} [copyright holder]

}
\keyword{internal}
