% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stac.R
\name{tidy_collections}
\alias{tidy_collections}
\title{Convert STAC collections to a tidy tibble}
\usage{
tidy_collections(collections)
}
\arguments{
\item{collections}{An rstac \code{doc_collections} object.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns: \code{id}, \code{title}, \code{description}.
}
\description{
Convert STAC collections to a tidy tibble
}
\keyword{internal}
